Deleted corva/state/ dir
