+++ src/corva/cache_adapter.py
@@ -1,4 +1,5 @@
-from typing import Optional, Protocol
+from datetime import timedelta
+from typing import Dict, List, Optional, Protocol, Union
 
 import redis
 
@@ -146,3 +147,112 @@ class RedisRepository:
 
     def vacuum(self, delete_count: int) -> None:
         self.lua_vacuum(keys=[self.hash_name, self.zset_name], args=[delete_count])
+
+
+class DeprecatedRedisAdapter:
+    """Expands basic redis functionality
+
+    Serves the purpose of adding custom logic on top of basic redis functions
+    (e.g. adding expiry to hset, which is not available out of the box).
+    """
+
+    DEFAULT_EXPIRY: timedelta = timedelta(days=60)
+    REDIS_STORED_VALUE_TYPE = Union[bytes, str, int, float]
+
+    def __init__(
+        self,
+        hash_name: str,
+        client: redis.Redis,
+    ):
+        self.default_name = hash_name
+        self.client = client
+
+        self.client.ping()
+
+    def hset(
+        self,
+        name: Optional[str] = None,
+        key: Optional[str] = None,
+        value: Optional[REDIS_STORED_VALUE_TYPE] = None,
+        mapping: Optional[Dict[str, REDIS_STORED_VALUE_TYPE]] = None,
+        expiry: Union[int, timedelta, None] = DEFAULT_EXPIRY,
+    ) -> int:
+        """Stores the data in cache
+
+        params:
+         key: key, which will contain the data
+         value: data to be saved
+         mapping: dict of key:data pairs to be saved
+         expiry: time in seconds for when data will be deleted from cache
+            expiration is reset with every hset call
+            expiration can be disabled by setting expiry to None
+        returns: number of inserted elements
+        """
+
+        name = name or self.default_name
+
+        n_set = self.client.hset(name=name, key=key, value=value, mapping=mapping)
+
+        if expiry is None and self.pttl(name=name) > 0:
+            self.client.persist(name=name)
+
+        if expiry is not None:
+            self.client.expire(name=name, time=expiry)
+
+        return n_set
+
+    def hget(
+        self, key: str, name: Optional[str] = None
+    ) -> Union[REDIS_STORED_VALUE_TYPE, None]:
+        """Loads data from cache
+
+        params:
+         key: key to load data from
+        returns: stored data
+        """
+
+        name = name or self.default_name
+        return self.client.hget(name=name, key=key)
+
+    def hgetall(
+        self, name: Optional[str] = None
+    ) -> Dict[str, Union[REDIS_STORED_VALUE_TYPE]]:
+        """Loads all data from cache"""
+
+        name = name or self.default_name
+        return self.client.hgetall(name=name)
+
+    def hdel(self, keys: List[str], name: Optional[str] = None) -> int:
+        """Deletes some data from cache
+
+        params:
+         keys: list of keys to delete
+        returns: number of deleted elements
+        """
+
+        name = name or self.default_name
+        return self.client.hdel(name, *keys)
+
+    def delete(self, *names: List[str]) -> int:
+        """Deletes all data from cache"""
+
+        names = names or [self.default_name]
+        return self.client.delete(*names)
+
+    def ttl(self, name: Optional[str] = None) -> int:
+        """Returns the number of seconds until expiration"""
+
+        name = name or self.default_name
+        return self.client.ttl(name=name)
+
+    def pttl(self, name: Optional[str] = None) -> int:
+        """Returns the number of milliseconds until expiration"""
+
+        name = name or self.default_name
+        return self.client.pttl(name=name)
+
+    def exists(self, *names: List[str]) -> int:
+        """Returns whether there is data in cache"""
+
+        names = names or [self.default_name]
+        return self.client.exists(*names)
